+++ lending/src/sdk/models/operations/uploadfiles.ts
@@ -3,16 +3,8 @@
  */
 
 import { SpeakeasyBase, SpeakeasyMetadata } from "../../../internal/utils";
-import * as shared from "../shared";
+import * as shared from "../../../sdk/models/shared";
 import { AxiosResponse } from "axios";
-
-export class UploadFilesRequestBody extends SpeakeasyBase {
-    @SpeakeasyMetadata({ data: "multipart_form, content=true" })
-    content: Uint8Array;
-
-    @SpeakeasyMetadata({ data: "multipart_form, name=requestBody" })
-    requestBody: string;
-}
 
 export class UploadFilesRequest extends SpeakeasyBase {
     /**
@@ -27,10 +19,8 @@ export class UploadFilesRequest extends SpeakeasyBase {
     @SpeakeasyMetadata({ data: "pathParam, style=simple;explode=false;name=connectionId" })
     connectionId: string;
 
-    @SpeakeasyMetadata({
-        data: "multipart_form, file=true, request, media_type=multipart/form-data",
-    })
-    requestBody?: UploadFilesRequestBody;
+    @SpeakeasyMetadata({ data: "request, media_type=multipart/form-data" })
+    fileUpload?: shared.FileUpload;
 }
 
 export class UploadFilesResponse extends SpeakeasyBase {
@@ -56,5 +46,5 @@ export class UploadFilesResponse extends SpeakeasyBase {
      * Raw HTTP response; suitable for custom response parsing
      */
     @SpeakeasyMetadata()
-    rawResponse?: AxiosResponse;
+    rawResponse: AxiosResponse;
 }