+++ lending/src/sdk/models/shared/accountingbill.ts
@@ -11,7 +11,7 @@ import { SupplementalData } from "./supplementaldata";
 import { SupplierRef } from "./supplierref";
 import { Expose, Type } from "class-transformer";
 
-export class AccountingBillPurchaseOrderReference extends SpeakeasyBase {
+export class PurchaseOrderReference extends SpeakeasyBase {
     /**
      * Identifier for the purchase order, unique for the company in the accounting platform.
      */
@@ -27,7 +27,7 @@ export class AccountingBillPurchaseOrderReference extends SpeakeasyBase {
     purchaseOrderNumber?: string;
 }
 
-export class AccountingBillWithholdingTax extends SpeakeasyBase {
+export class WithholdingTax extends SpeakeasyBase {
     /**
      * Amount of tax withheld.
      */
@@ -117,6 +117,13 @@ export class AccountingBill extends SpeakeasyBase {
      * | **GBP**          | £20            | 1.277         | $25.54                     |
      * | **EUR**          | €20            | 1.134         | $22.68                     |
      * | **RUB**          | ₽20            | 0.015         | $0.30                      |
+     *
+     *
+     * ### Integration-specific details
+     *
+     * | Integration       | Scenario                                        | System behavior                                                                                                                                                      |
+     * |-------------------|-------------------------------------------------|----------------------------------------------------------------------------------------------------------------------------------------------------------------------|
+     * | QuickBooks Online | Transaction currency differs from base currency | If currency rate value is left `null`, a rate of 1 will be used by QBO by default. To override this, include the required currency rate in the expense transaction.  |
      */
     @SpeakeasyMetadata()
     @Expose({ name: "currencyRate" })
@@ -169,10 +176,10 @@ export class AccountingBill extends SpeakeasyBase {
     @Type(() => AccountingPaymentAllocation)
     paymentAllocations?: AccountingPaymentAllocation[];
 
-    @SpeakeasyMetadata({ elemType: AccountingBillPurchaseOrderReference })
+    @SpeakeasyMetadata({ elemType: PurchaseOrderReference })
     @Expose({ name: "purchaseOrderRefs" })
-    @Type(() => AccountingBillPurchaseOrderReference)
-    purchaseOrderRefs?: AccountingBillPurchaseOrderReference[];
+    @Type(() => PurchaseOrderReference)
+    purchaseOrderRefs?: PurchaseOrderReference[];
 
     /**
      * User-friendly reference for the bill.
@@ -233,8 +240,8 @@ export class AccountingBill extends SpeakeasyBase {
     @Expose({ name: "totalAmount" })
     totalAmount: number;
 
-    @SpeakeasyMetadata({ elemType: AccountingBillWithholdingTax })
+    @SpeakeasyMetadata({ elemType: WithholdingTax })
     @Expose({ name: "withholdingTax" })
-    @Type(() => AccountingBillWithholdingTax)
-    withholdingTax?: AccountingBillWithholdingTax[];
+    @Type(() => WithholdingTax)
+    withholdingTax?: WithholdingTax[];
 }