+++ lending/src/sdk/models/shared/accountingpaymentallocation.ts
@@ -6,7 +6,7 @@ import { SpeakeasyBase, SpeakeasyMetadata } from "../../../internal/utils";
 import { PaymentAllocationPayment } from "./paymentallocationpayment";
 import { Expose, Type } from "class-transformer";
 
-export class AccountingPaymentAllocationAllocation extends SpeakeasyBase {
+export class Allocation extends SpeakeasyBase {
     /**
      * In Codat's data model, dates and times are represented using the <a class="external" href="https://en.wikipedia.org/wiki/ISO_8601" target="_blank">ISO 8601 standard</a>. Date and time fields are formatted as strings; for example:
      *
@@ -77,6 +77,13 @@ export class AccountingPaymentAllocationAllocation extends SpeakeasyBase {
      * | **GBP**          | £20            | 1.277         | $25.54                     |
      * | **EUR**          | €20            | 1.134         | $22.68                     |
      * | **RUB**          | ₽20            | 0.015         | $0.30                      |
+     *
+     *
+     * ### Integration-specific details
+     *
+     * | Integration       | Scenario                                        | System behavior                                                                                                                                                      |
+     * |-------------------|-------------------------------------------------|----------------------------------------------------------------------------------------------------------------------------------------------------------------------|
+     * | QuickBooks Online | Transaction currency differs from base currency | If currency rate value is left `null`, a rate of 1 will be used by QBO by default. To override this, include the required currency rate in the expense transaction.  |
      */
     @SpeakeasyMetadata()
     @Expose({ name: "currencyRate" })
@@ -93,8 +100,8 @@ export class AccountingPaymentAllocationAllocation extends SpeakeasyBase {
 export class AccountingPaymentAllocation extends SpeakeasyBase {
     @SpeakeasyMetadata()
     @Expose({ name: "allocation" })
-    @Type(() => AccountingPaymentAllocationAllocation)
-    allocation: AccountingPaymentAllocationAllocation;
+    @Type(() => Allocation)
+    allocation: Allocation;
 
     @SpeakeasyMetadata()
     @Expose({ name: "payment" })