+++ lending/src/sdk/models/shared/billcreditnotelineitem.ts
@@ -6,12 +6,13 @@ import { SpeakeasyBase, SpeakeasyMetadata } from "../../../internal/utils";
 import { AccountRef } from "./accountref";
 import { AccountsPayableTracking } from "./accountspayabletracking";
 import { TrackingCategoryRef } from "./trackingcategoryref";
+import { Zero } from "./zero";
 import { Expose, Type } from "class-transformer";
 
 /**
  * Reference to the item the line is linked to.
  */
-export class BillCreditNoteLineItemItemReference extends SpeakeasyBase {
+export class ItemReference extends SpeakeasyBase {
     /**
      * Unique identifier for the item in the accounting platform.
      */
@@ -41,7 +42,7 @@ export class BillCreditNoteLineItemItemReference extends SpeakeasyBase {
  * - Invoice line items
  * - Items
  */
-export class BillCreditNoteLineItemTaxRateReference extends SpeakeasyBase {
+export class TaxRateReference extends SpeakeasyBase {
     /**
      * Applicable tax rate.
      */
@@ -74,6 +75,18 @@ export class BillCreditNoteLineItem extends SpeakeasyBase {
     accountRef?: AccountRef;
 
     /**
+     * Links the current record line to the underlying record line that created it.
+     *
+     * @remarks
+     *
+     * For example, if a bill is generated from a purchase order, this property allows you to connect the bill line item to the purchase order line item in our data model.
+     */
+    @SpeakeasyMetadata()
+    @Expose({ name: "createdFromLineRef" })
+    @Type(() => Zero)
+    createdFromLineRef?: Zero;
+
+    /**
      * Friendly name of each line item. For example, the goods or service for which credit has been received.
      */
     @SpeakeasyMetadata()
@@ -99,8 +112,8 @@ export class BillCreditNoteLineItem extends SpeakeasyBase {
      */
     @SpeakeasyMetadata()
     @Expose({ name: "itemRef" })
-    @Type(() => BillCreditNoteLineItemItemReference)
-    itemRef?: BillCreditNoteLineItemItemReference;
+    @Type(() => ItemReference)
+    itemRef?: ItemReference;
 
     /**
      * Number of units of the goods or service for which credit has been received.
@@ -139,8 +152,8 @@ export class BillCreditNoteLineItem extends SpeakeasyBase {
      */
     @SpeakeasyMetadata()
     @Expose({ name: "taxRateRef" })
-    @Type(() => BillCreditNoteLineItemTaxRateReference)
-    taxRateRef?: BillCreditNoteLineItemTaxRateReference;
+    @Type(() => TaxRateReference)
+    taxRateRef?: TaxRateReference;
 
     /**
      * Total amount of the line item, including discounts and tax.
@@ -173,4 +186,11 @@ export class BillCreditNoteLineItem extends SpeakeasyBase {
     @SpeakeasyMetadata()
     @Expose({ name: "unitAmount" })
     unitAmount: number;
+
+    /**
+     * The measurement which defines a unit for this item (e.g. 'kilogram', 'litre').
+     */
+    @SpeakeasyMetadata()
+    @Expose({ name: "unitOfMeasurement" })
+    unitOfMeasurement?: string;
 }