+++ lending/RELEASES.md
@@ -9,3 +9,173 @@ Based on:
 - [typescript v0.1.0] lending
 ### Releases
 - [NPM v0.1.0] https://www.npmjs.com/package/openapi/v/0.1.0 - lending
+
+## 2023-11-06 01:55:50
+### Changes
+Based on:
+- OpenAPI Doc 3.0.0 https://raw.githubusercontent.com/codatio/oas/main/yaml/Codat-Lending.yaml
+- Speakeasy CLI 1.112.1 (2.173.0) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v0.2.0] lending
+### Releases
+- [NPM v0.2.0] https://www.npmjs.com/package/lending/v/0.2.0 - lending
+
+## 2023-11-13 01:56:09
+### Changes
+Based on:
+- OpenAPI Doc 3.0.0 https://raw.githubusercontent.com/codatio/oas/main/yaml/Codat-Lending.yaml
+- Speakeasy CLI 1.119.0 (2.188.1) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v0.3.0] lending
+### Releases
+- [NPM v0.3.0] https://www.npmjs.com/package/lending/v/0.3.0 - lending
+
+## 2023-11-20 01:58:00
+### Changes
+Based on:
+- OpenAPI Doc 3.0.0 https://raw.githubusercontent.com/codatio/oas/main/yaml/Codat-Lending.yaml
+- Speakeasy CLI 1.121.1 (2.194.1) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v0.3.1] lending
+### Releases
+- [NPM v0.3.1] https://www.npmjs.com/package/lending/v/0.3.1 - lending
+
+## 2023-11-27 01:57:23
+### Changes
+Based on:
+- OpenAPI Doc 3.0.0 https://raw.githubusercontent.com/codatio/oas/main/yaml/Codat-Lending.yaml
+- Speakeasy CLI 1.122.0 (2.195.2) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v0.3.2] lending
+### Releases
+- [NPM v0.3.2] https://www.npmjs.com/package/lending/v/0.3.2 - lending
+
+## 2023-12-04 01:58:52
+### Changes
+Based on:
+- OpenAPI Doc 3.0.0 https://raw.githubusercontent.com/codatio/oas/main/yaml/Codat-Lending.yaml
+- Speakeasy CLI 1.125.0 (2.209.0) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v0.3.3] lending
+### Releases
+- [NPM v0.3.3] https://www.npmjs.com/package/lending/v/0.3.3 - lending
+
+## 2023-12-11 01:58:32
+### Changes
+Based on:
+- OpenAPI Doc 3.0.0 https://raw.githubusercontent.com/codatio/oas/main/yaml/Codat-Lending.yaml
+- Speakeasy CLI 1.125.2 (2.210.6) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v0.3.4] lending
+### Releases
+- [NPM v0.3.4] https://www.npmjs.com/package/lending/v/0.3.4 - lending
+
+## 2023-12-18 01:58:09
+### Changes
+Based on:
+- OpenAPI Doc 3.0.0 https://raw.githubusercontent.com/codatio/oas/main/yaml/Codat-Lending.yaml
+- Speakeasy CLI 1.126.4 (2.214.10) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v0.4.0] lending
+### Releases
+- [NPM v0.4.0] https://www.npmjs.com/package/lending/v/0.4.0 - lending
+
+## 2023-12-25 01:29:55
+### Changes
+Based on:
+- OpenAPI Doc 3.0.0 https://raw.githubusercontent.com/codatio/oas/main/yaml/Codat-Lending.yaml
+- Speakeasy CLI 1.129.1 (2.223.3) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v0.4.1] lending
+### Releases
+- [NPM v0.4.1] https://www.npmjs.com/package/lending/v/0.4.1 - lending
+
+## 2024-01-08 01:30:22
+### Changes
+Based on:
+- OpenAPI Doc 3.0.0 https://raw.githubusercontent.com/codatio/oas/main/yaml/Codat-Lending.yaml
+- Speakeasy CLI 1.133.1 (2.228.1) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v0.5.0] lending
+### Releases
+- [NPM v0.5.0] https://www.npmjs.com/package/lending/v/0.5.0 - lending
+
+## 2024-01-15 01:31:46
+### Changes
+Based on:
+- OpenAPI Doc 3.0.0 https://raw.githubusercontent.com/codatio/oas/main/yaml/Codat-Lending.yaml
+- Speakeasy CLI 1.140.0 (2.231.0) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v0.5.1] lending
+### Releases
+- [NPM v0.5.1] https://www.npmjs.com/package/lending/v/0.5.1 - lending
+
+## 2024-02-05 01:28:07
+### Changes
+Based on:
+- OpenAPI Doc 3.0.0 https://raw.githubusercontent.com/codatio/oas/main/yaml/Codat-Lending.yaml
+- Speakeasy CLI 1.166.0 (2.248.6) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v0.6.0] lending
+### Releases
+- [NPM v0.6.0] https://www.npmjs.com/package/lending/v/0.6.0 - lending
+
+## 2024-02-19 01:27:02
+### Changes
+Based on:
+- OpenAPI Doc 3.0.0 https://raw.githubusercontent.com/codatio/oas/main/yaml/Codat-Lending.yaml
+- Speakeasy CLI 1.183.2 (2.262.2) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v0.6.1] lending
+### Releases
+- [NPM v0.6.1] https://www.npmjs.com/package/lending/v/0.6.1 - lending
+
+## 2024-02-26 01:27:06
+### Changes
+Based on:
+- OpenAPI Doc 3.0.0 https://raw.githubusercontent.com/codatio/oas/main/yaml/Codat-Lending.yaml
+- Speakeasy CLI 1.193.4 (2.272.4) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v0.6.2] lending
+### Releases
+- [NPM v0.6.2] https://www.npmjs.com/package/lending/v/0.6.2 - lending
+
+## 2024-03-04 01:28:27
+### Changes
+Based on:
+- OpenAPI Doc 3.0.0 https://raw.githubusercontent.com/codatio/oas/main/yaml/Codat-Lending.yaml
+- Speakeasy CLI 1.200.0 (2.277.0) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v0.7.0] lending
+### Releases
+- [NPM v0.7.0] https://www.npmjs.com/package/lending/v/0.7.0 - lending
+
+## 2024-03-11 01:26:29
+### Changes
+Based on:
+- OpenAPI Doc 3.0.0 https://raw.githubusercontent.com/codatio/oas/main/yaml/Codat-Lending.yaml
+- Speakeasy CLI 1.205.0 (2.279.1) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v0.7.1] lending
+### Releases
+- [NPM v0.7.1] https://www.npmjs.com/package/lending/v/0.7.1 - lending
+
+## 2024-03-18 01:26:35
+### Changes
+Based on:
+- OpenAPI Doc 3.0.0 https://raw.githubusercontent.com/codatio/oas/main/yaml/Codat-Lending.yaml
+- Speakeasy CLI 1.209.3 (2.281.2) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v0.7.2] lending
+### Releases
+- [NPM v0.7.2] https://www.npmjs.com/package/lending/v/0.7.2 - lending
+
+## 2024-03-25 01:52:44
+### Changes
+Based on:
+- OpenAPI Doc 3.0.0 https://raw.githubusercontent.com/codatio/oas/main/yaml/Codat-Lending.yaml
+- Speakeasy CLI 1.221.0 (2.287.0) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v0.7.3] lending
+### Releases
+- [NPM v0.7.3] https://www.npmjs.com/package/lending/v/0.7.3 - lending